# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Febrafar - Planner",
        "description": "Febrafar planner",
        "version": "1.0"
    },
    "paths": {
...
        "/api/users/matches": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Find all users who match specific criteria",
                "operationId": "de1376c8128aa20cd61a6516150008d1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name to be founded",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve the list of users",
                "operationId": "271a9018597bbdb4204e778b4f2db136",
                "responses": {
                    "200": {
                        "description": "List of all users successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "a794d2c5818707a660863ff972b607ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "New user name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "New user email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "New user password",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "New user role",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Recover a single user",
                "operationId": "64c59785aec5f2658a4a2b5a560cb070",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User recovered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update an existing user",
                "operationId": "f1f65c5e73cb0b10fdd850447b672c28",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "New username",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "New user email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "New user password",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "New user role",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "422": {
                        "description": "Dados de entrada inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete an existing user",
                "operationId": "e61da8b22508c63df70e810437466d03",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "description": "user id",
                        "type": "string"
                    },
                    "name": {
                        "description": "user name",
                        "type": "string"
                    },
                    "email": {
                        "description": "user email",
                        "type": "string"
                    },
                    "password": {
                        "description": "user password",
                        "type": "string"
                    },
                    "role": {
                        "description": "user role",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}